Move key-text extraction to a util class
